Get rid of more warnings.  (Hope they're all correct...)

+++ sf-cvs/trunk/audacity-src/src/Internat.cpp
@@ -289,7 +289,7 @@ wxString Internat::StripAccelerators(const wxString &s)
 {
    wxString result;
    result.Alloc(s.Length());
-   for(int i = 0; i < s.Length(); i++) {
+   for(size_t i = 0; i < s.Length(); i++) {
       if (s[i] == '\t')
          break;
       if (s[i] != '&' && s[i] != '.')
